Moved the ApexLegendsAPI class to its own file in the package root
Resolves #12
Resolves #3

- the module is now exporting all the classes through the `__init__.py` which will make importing
 simpler
- Moved `ApexLegendsAPI` class to its own file

- Removed debug code that dumped the response json to a file
