+++ setup.py
@@ -8,7 +8,7 @@ with open('requirements.txt') as f:
 
 setup(
     name='apex-legends-api',
-    version='0.2.0',
+    version='0.2.1',
     packages=['apex_legends_api'],
     python_requires='>=3.6.*',
     url='https://github.com/johnsturgeon/apex-legends-api',
